Separate HWCommandQueue in it's own header
